+++ packages/common/src/directives/ng_class.ts
@@ -5,69 +5,9 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {Directive, DoCheck, Input, ɵRenderFlags, ɵɵclassMap, ɵɵdefineDirective} from '@angular/core';
+import {Directive, DoCheck, ElementRef, Input, IterableChanges, IterableDiffer, IterableDiffers, KeyValueChanges, KeyValueDiffer, KeyValueDiffers, Renderer2, ɵisListLikeIterable as isListLikeIterable, ɵstringify as stringify} from '@angular/core';
 
-import {NgClassImpl, NgClassImplProvider} from './ng_class_impl';
-
-
-
-/*
- * NgClass (as well as NgStyle) behaves differently when loaded in the VE and when not.
- *
- * If the VE is present (which is for older versions of Angular) then NgClass will inject
- * the legacy diffing algorithm as a service and delegate all styling changes to that.
- *
- * If the VE is not present then NgStyle will normalize (through the injected service) and
- * then write all styling changes to the `[style]` binding directly (through a host binding).
- * Then Angular will notice the host binding change and treat the changes as styling
- * changes and apply them via the core styling instructions that exist within Angular.
- */
-
-// used when the VE is present
-export const ngClassDirectiveDef__PRE_R3__ = undefined;
-
-// used when the VE is not present (note the directive will
-// never be instantiated normally because it is apart of a
-// base class)
-export const ngClassDirectiveDef__POST_R3__ = ɵɵdefineDirective({
-  type: function() {} as any,
-  selectors: null as any,
-  hostVars: 2,
-  hostBindings: function(rf: ɵRenderFlags, ctx: any, elIndex: number) {
-    if (rf & ɵRenderFlags.Update) {
-      ɵɵclassMap(ctx.getValue());
-    }
-  }
-});
-
-export const ngClassDirectiveDef = ngClassDirectiveDef__PRE_R3__;
-
-export const ngClassFactoryDef__PRE_R3__ = undefined;
-export const ngClassFactoryDef__POST_R3__ = function() {};
-export const ngClassFactoryDef = ngClassFactoryDef__PRE_R3__;
-
-/**
- * Serves as the base non-VE container for NgClass.
- *
- * While this is a base class that NgClass extends from, the
- * class itself acts as a container for non-VE code to setup
- * a link to the `[class]` host binding (via the static
- * `ɵdir` property on the class).
- *
- * Note that the `ɵdir` property's code is switched
- * depending if VE is present or not (this allows for the
- * binding code to be set only for newer versions of Angular).
- *
- * @publicApi
- */
-export class NgClassBase {
-  static ɵdir: any = ngClassDirectiveDef;
-  static ɵfac: any = ngClassFactoryDef;
-
-  constructor(protected _delegate: NgClassImpl) {}
-
-  getValue() { return this._delegate.getValue(); }
-}
+type NgClassSupportedTypes = string[] | Set<string>| {[klass: string]: any} | null | undefined;
 
 /**
  * @ngModule CommonModule
@@ -97,17 +37,124 @@ export class NgClassBase {
  *
  * @publicApi
  */
-@Directive({selector: '[ngClass]', providers: [NgClassImplProvider]})
-export class NgClass extends NgClassBase implements DoCheck {
-  constructor(delegate: NgClassImpl) { super(delegate); }
+@Directive({selector: '[ngClass]'})
+export class NgClass implements DoCheck {
+  private _iterableDiffer: IterableDiffer<string>|null = null;
+  private _keyValueDiffer: KeyValueDiffer<string, any>|null = null;
+  private _initialClasses: string[] = [];
+  private _rawClass: NgClassSupportedTypes = null;
+
+  constructor(
+      private _iterableDiffers: IterableDiffers, private _keyValueDiffers: KeyValueDiffers,
+      private _ngEl: ElementRef, private _renderer: Renderer2) {}
+
 
   @Input('class')
-  set klass(value: string) { this._delegate.setClass(value); }
+  set klass(value: string) {
+    this._removeClasses(this._initialClasses);
+    this._initialClasses = typeof value === 'string' ? value.split(/\s+/) : [];
+    this._applyClasses(this._initialClasses);
+    this._applyClasses(this._rawClass);
+  }
 
   @Input('ngClass')
   set ngClass(value: string|string[]|Set<string>|{[klass: string]: any}) {
-    this._delegate.setNgClass(value);
+    this._removeClasses(this._rawClass);
+    this._applyClasses(this._initialClasses);
+
+    this._iterableDiffer = null;
+    this._keyValueDiffer = null;
+
+    this._rawClass = typeof value === 'string' ? value.split(/\s+/) : value;
+
+    if (this._rawClass) {
+      if (isListLikeIterable(this._rawClass)) {
+        this._iterableDiffer = this._iterableDiffers.find(this._rawClass).create();
+      } else {
+        this._keyValueDiffer = this._keyValueDiffers.find(this._rawClass).create();
+      }
+    }
   }
 
-  ngDoCheck() { this._delegate.applyChanges(); }
+  ngDoCheck() {
+    if (this._iterableDiffer) {
+      const iterableChanges = this._iterableDiffer.diff(this._rawClass as string[]);
+      if (iterableChanges) {
+        this._applyIterableChanges(iterableChanges);
+      }
+    } else if (this._keyValueDiffer) {
+      const keyValueChanges = this._keyValueDiffer.diff(this._rawClass as{[k: string]: any});
+      if (keyValueChanges) {
+        this._applyKeyValueChanges(keyValueChanges);
+      }
+    }
+  }
+
+  private _applyKeyValueChanges(changes: KeyValueChanges<string, any>): void {
+    changes.forEachAddedItem((record) => this._toggleClass(record.key, record.currentValue));
+    changes.forEachChangedItem((record) => this._toggleClass(record.key, record.currentValue));
+    changes.forEachRemovedItem((record) => {
+      if (record.previousValue) {
+        this._toggleClass(record.key, false);
+      }
+    });
+  }
+
+  private _applyIterableChanges(changes: IterableChanges<string>): void {
+    changes.forEachAddedItem((record) => {
+      if (typeof record.item === 'string') {
+        this._toggleClass(record.item, true);
+      } else {
+        throw new Error(
+            `NgClass can only toggle CSS classes expressed as strings, got ${stringify(record.item)}`);
+      }
+    });
+
+    changes.forEachRemovedItem((record) => this._toggleClass(record.item, false));
+  }
+
+  /**
+   * Applies a collection of CSS classes to the DOM element.
+   *
+   * For argument of type Set and Array CSS class names contained in those collections are always
+   * added.
+   * For argument of type Map CSS class name in the map's key is toggled based on the value (added
+   * for truthy and removed for falsy).
+   */
+  private _applyClasses(rawClassVal: NgClassSupportedTypes) {
+    if (rawClassVal) {
+      if (Array.isArray(rawClassVal) || rawClassVal instanceof Set) {
+        (<any>rawClassVal).forEach((klass: string) => this._toggleClass(klass, true));
+      } else {
+        Object.keys(rawClassVal).forEach(klass => this._toggleClass(klass, !!rawClassVal[klass]));
+      }
+    }
+  }
+
+  /**
+   * Removes a collection of CSS classes from the DOM element. This is mostly useful for cleanup
+   * purposes.
+   */
+  private _removeClasses(rawClassVal: NgClassSupportedTypes) {
+    if (rawClassVal) {
+      if (Array.isArray(rawClassVal) || rawClassVal instanceof Set) {
+        (<any>rawClassVal).forEach((klass: string) => this._toggleClass(klass, false));
+      } else {
+        Object.keys(rawClassVal).forEach(klass => this._toggleClass(klass, false));
+      }
+    }
+  }
+
+  private _toggleClass(klass: string, enabled: boolean): void {
+    klass = klass.trim();
+    if (klass) {
+      klass.split(/\s+/g).forEach(klass => {
+        if (enabled) {
+          this._renderer.addClass(this._ngEl.nativeElement, klass);
+        } else {
+          this._renderer.removeClass(this._ngEl.nativeElement, klass);
+        }
+      });
+    }
+  }
 }
